refactor(auth): rm unsued useLogout & refactor
