+++ src/features/auth/context/AuthContext.tsx
@@ -4,7 +4,8 @@ import { getErrorMessage } from '../../../lib/errors/errors-utils';
 import { authUtils } from '../utils/auth-utils';
 import { useSignin } from '../hooks/mutations/useSignin';
 import { useSignup } from '../hooks/mutations/useSignup';
-import { useCurrentUser, userKeys } from '../hooks/queries/useCurrentUser';
+import { useCurrentUser } from '../hooks/queries/useCurrentUser';
+import { userKeys } from '../utils/authQueryKeys';
 import { AuthErrorCode, User } from '../types';
 import { ApiError } from '../../../lib/client/types';
 import { useQueryClient } from '@tanstack/react-query';
