+++ src/features/auth/utils/authQueryKeys.ts
@@ -1,0 +1,4 @@
+export const userKeys = {
+  all: ['user'] as const,
+  currentUser: () => [...userKeys.all, 'current'] as const,
+};
